requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-allauth==65.0.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==24.1
psycopg2==2.9.10
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
